+++ src/main/java/com/nearby/shops/controller/CarteBancaireController.java
@@ -24,7 +24,7 @@ public class CarteBancaireController {
     }
 
     @GetMapping("/{id}")
-    public Optional<CarteBancaire> GetByID(@PathVariable Integer id) {return this.carteBancaireService.FindByID(id);}
+    public Optional<CarteBancaire> GetByID(@PathVariable Long id) {return this.carteBancaireService.FindByID(id);}
 
     @PostMapping
     public CarteBancaire Add(@RequestBody CarteBancaire nouvelleCarteBancaire) {
@@ -32,12 +32,12 @@ public class CarteBancaireController {
     }
 
     @DeleteMapping("/{id}")
-    public void DeleteById(@PathVariable Integer id) {
+    public void DeleteById(@PathVariable Long id) {
         this.carteBancaireService.DeleteById(id);
     }
 
     @PutMapping("/{id}")
-    CarteBancaire Update(@RequestBody CarteBancaire nouvelleCarteBancaire, @PathVariable Integer id) {
+    CarteBancaire Update(@RequestBody CarteBancaire nouvelleCarteBancaire, @PathVariable Long id) {
         return this.carteBancaireService.Update(nouvelleCarteBancaire,id);
     }
 }
