+++ src/main/java/com/nearby/shops/dao/CrudDao.java
@@ -6,10 +6,10 @@ import java.util.List;
 import java.util.Optional;
 
 public interface CrudDao<T> {
-    public Optional<T> FindByID(Integer arg);
+    public Optional<T> FindByID(Long arg);
     public T Add(T arg);
-    public void DeleteById(Integer arg);
-    public T Update(T obj,Integer arg);
+    public void DeleteById(Long arg);
+    public T Update(T obj,Long arg);
     public List<T> GetAll();
 
 }
