+++ src/main/java/com/nearby/shops/repositories/BoutiqueRepository.java
@@ -3,5 +3,5 @@ package com.nearby.shops.repositories;
 import com.nearby.shops.models.Boutique;
 import org.springframework.data.repository.CrudRepository;
 
-public interface BoutiqueRepository extends CrudRepository<Boutique,Integer> {
+public interface BoutiqueRepository extends CrudRepository<Boutique,Long> {
 }
