+++ src/main/java/com/nearby/shops/repositories/CarteBancaireRepository.java
@@ -3,5 +3,5 @@ package com.nearby.shops.repositories;
 import com.nearby.shops.models.CarteBancaire;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CarteBancaireRepository extends CrudRepository<CarteBancaire,Integer> {
+public interface CarteBancaireRepository extends CrudRepository<CarteBancaire,Long> {
 }
