+++ src/main/java/com/nearby/shops/repositories/CashOnDeliveryRepository.java
@@ -3,5 +3,5 @@ package com.nearby.shops.repositories;
 import com.nearby.shops.models.CashOnDelivery;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CashOnDeliveryRepository extends CrudRepository<CashOnDelivery,Integer> {
+public interface CashOnDeliveryRepository extends CrudRepository<CashOnDelivery,Long> {
 }
