+++ src/main/java/com/nearby/shops/repositories/CategorieRepository.java
@@ -3,6 +3,6 @@ package com.nearby.shops.repositories;
 import com.nearby.shops.models.Categorie;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CategorieRepository extends CrudRepository<Categorie,Integer>{
+public interface CategorieRepository extends CrudRepository<Categorie,Long>{
 
 }
