+++ src/main/java/com/nearby/shops/repositories/CommandeRepository.java
@@ -3,5 +3,5 @@ package com.nearby.shops.repositories;
 import com.nearby.shops.models.Commande;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CommandeRepository extends CrudRepository<Commande,Integer> {
+public interface CommandeRepository extends CrudRepository<Commande,Long> {
 }
