+++ src/main/java/com/nearby/shops/repositories/LivraisonRepository.java
@@ -3,5 +3,5 @@ package com.nearby.shops.repositories;
 import com.nearby.shops.models.Livraison;
 import org.springframework.data.repository.CrudRepository;
 
-public interface LivraisonRepository extends CrudRepository<Livraison,Integer> {
+public interface LivraisonRepository extends CrudRepository<Livraison,Long> {
 }
