+++ src/main/java/com/nearby/shops/repositories/PaymentRepository.java
@@ -3,5 +3,5 @@ package com.nearby.shops.repositories;
 import com.nearby.shops.models.Payment;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PaymentRepository extends CrudRepository<Payment, Integer> {
+public interface PaymentRepository extends CrudRepository<Payment, Long> {
 }
