+++ src/main/java/com/nearby/shops/repositories/PersonneRepository.java
@@ -3,6 +3,6 @@ package com.nearby.shops.repositories;
 import com.nearby.shops.models.Personne;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PersonneRepository extends CrudRepository<Personne,Integer> {
+public interface PersonneRepository extends CrudRepository<Personne,Long> {
 
 }
