+++ src/main/java/com/nearby/shops/repositories/ProduitRepository.java
@@ -3,5 +3,5 @@ package com.nearby.shops.repositories;
 import com.nearby.shops.models.Produit;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProduitRepository extends CrudRepository<Produit,Integer>{
+public interface ProduitRepository extends CrudRepository<Produit,Long>{
 }
