+++ src/main/java/com/nearby/shops/repositories/VilleRepository.java
@@ -2,5 +2,5 @@ package com.nearby.shops.repositories;
 import com.nearby.shops.models.Ville;
 import org.springframework.data.repository.CrudRepository;
 
-public interface VilleRepository extends CrudRepository<Ville,Integer> {
+public interface VilleRepository extends CrudRepository<Ville,Long> {
 }
